\name{cddat}
\alias{cddat}
\title{Old functions from NOISeq}
\usage{
    cddat(input)
}
\arguments{
    \item{input}{input to cddat.}
}
\value{
    a list with data to plot.
}
\description{
    Old functions from NOISeq to create the
    \code{"readnoise"} plots. Internal use only.
}
\note{
    Adopted from an older version of NOISeq package (author:
    Sonia Tarazona).
}
\author{
    Panagiotis Moulos
}

