\name{cdplot}
\alias{cdplot}
\title{Old functions from NOISeq}
\usage{
    cdplot(dat, samples = NULL, ...)
}
\arguments{
    \item{dat}{the returned list from \code{\link{cddat}}.}

    \item{samples}{the samples to plot.}

    \item{...}{further arguments passed to e.g.
    \code{\link{par}}.}
}
\value{
    Nothing, it created the old RNA composition plot.
}
\description{
    Old functions from NOISeq to create the
    \code{"readnoise"} plots. Internal use only.
}
\note{
    Adopted from an older version of NOISeq package (author:
    Sonia Tarazona)
}
\author{
    Panagiotis Moulos
}

