\name{disp}
\alias{disp}
\title{Message displayer}
\usage{
    disp(...)
}
\arguments{
    \item{...}{a vector of elements that compose the display
    message.}
}
\description{
    Displays a message during execution of the several
    functions. Internal use.
}
\examples{
i <- 1
disp("Now running iteration ",i,"...")
}
\author{
    Panagiotis Moulos
}

