\name{graphics.close}
\alias{graphics.close}
\title{Close plotting device}
\usage{
    graphics.close(o)
}
\arguments{
    \item{o}{the plotting device, see main metaseqr function}
}
\description{
    Wrapper function to close a plotting device. Internal use
    only.
}
\examples{
\donttest{
graphics.close("pdf")
}
}
\author{
    Panagiotis Moulos
}

