\docType{data}
\name{hg19.exon.counts}
\alias{hg19.exon.counts}
\title{Human RNA-Seq data with three conditions, three samples}
\format{a \code{data.frame} with exon read counts and some embedded annotation, one row per exon.}
\source{
    GEO (http://www.ncbi.nlm.nih.gov/geo/)
}
\description{
    This data set contains RNA-Seq exon read counts for 3
    chromosomes. The data are from an experiment studying the
    effect of a long non-coding RNA related to the ASCL2 gene
    in WNT signaling and intestinal cancer. It has two
    conditions (CON, DOX) and four samples (CON_BR1, CON_BR2,
    DOX_BR1, DOX_BR2). It also contains a predefined
    \code{sample.list} and \code{libsize.list} named
    \code{sample.list.hg18} and \code{libsize.list.hg18}.
}
\author{
    Panagiotis Moulos
}
\keyword{datasets}

