\docType{data}
\name{libsize.list.hg19}
\alias{libsize.list.hg19}
\title{Human RNA-Seq data with three conditions, three samples}
\format{a named \code{list} with library sizes.}
\source{
    GEO (http://www.ncbi.nlm.nih.gov/geo/)
}
\description{
    The library size list for \code{hg19.exon.counts}. See
    the data set description.
}
\author{
    Panagiotis Moulos
}
\keyword{datasets}

