\name{log2disp}
\alias{log2disp}
\title{Display value transformation}
\usage{
    log2disp(mat, base = 2)
}
\arguments{
    \item{mat}{input data matrix}

    \item{base}{logarithmic base, 2 or 10}
}
\description{
    Logarithmic transformation for display purposes. Internal
    use only.
}
\author{
    Panagiotis Moulos
}

