\name{make.export.list}
\alias{make.export.list}
\title{Intitialize output list}
\usage{
    make.export.list(con)
}
\arguments{
    \item{con}{The contrasts.}
}
\value{
    An empty named list.
}
\description{
    Initializes metaseqr R output. Internal use only.
}
\author{
    Panagiotis Moulos
}

