\name{make.project.path}
\alias{make.project.path}
\title{Project path constructor}
\usage{
    make.project.path(path, f = NULL)
}
\arguments{
    \item{path}{The desired project path. Can be NULL for
    auto-generation.}

    \item{f}{The input counts table file.}
}
\value{
    A list with project path elements.
}
\description{
    Create the main metaseqr project path. Internal use only.
}
\author{
    Panagiotis Moulos
}

