\name{make.report.messages}
\alias{make.report.messages}
\title{Initializer of report messages}
\usage{
    make.report.messages(lang)
}
\arguments{
    \item{lang}{The language of the report. For now, only
    english (\code{"en"}) is supported.}
}
\value{
    An named list with messages for each input option.
}
\description{
    Initializes metaseqR report tmeplate messages output.
    Internal use only.
}
\author{
    Panagiotis Moulos
}

