\name{nat2log}
\alias{nat2log}
\title{General value transformation}
\usage{
    nat2log(x, base = 2, off = 1)
}
\arguments{
    \item{x}{input data matrix}

    \item{base}{logarithmic base, 2 or 10}

    \item{off}{offset to avoid Infinity}
}
\description{
    Logarithmic transformation. Internal use only.
}
\author{
    Panagiotis Moulos
}

