\docType{data}
\name{sample.list.hg19}
\alias{sample.list.hg19}
\title{Human RNA-Seq data with three conditions, three samples}
\format{a named \code{list} with condition and sample names.}
\source{
    GEO (http://www.ncbi.nlm.nih.gov/geo/)
}
\description{
    The sample list for \code{hg19.exon.counts}. See the data
    set description.
}
\author{
    Panagiotis Moulos
}
\keyword{datasets}

