\docType{data}
\name{sample.list.mm9}
\alias{sample.list.mm9}
\title{Mouse RNA-Seq data with two conditions, four samples}
\format{a named \code{list} with condition and sample names.}
\source{
    ENCODE (http://genome.ucsc.edu/encode/)
}
\description{
    The sample list for \code{mm9.gene.counts}. See the data
    set description.
}
\author{
    Panagiotis Moulos
}
\keyword{datasets}

